requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.0
httpx==0.27.2
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
pydantic==2.9.2
pydantic-settings==2.5.2
